requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.0.1
idna==2.8
idna-ssl==1.1.0
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.5.2
pylint==2.3.1
python-dotenv==0.10.1
six==1.12.0
typed-ast==1.3.1
typing-extensions==3.7.2
websockets==6.0
wrapt==1.11.1
yarl==1.3.0
